+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 * Add support for automatic language injection on the minted environment
+* Add support for DeclareGraphicsExtensions
+* Add inspection to warn about a missing reference for a glossary occurrence
 * Do not fold sections in a command definition
 * Include optional parameters in spellcheck, if it contains text